fix(observe): capture Claude stream usage
